OLE-8027 Prorated Additional Charges Associated with First PO on Invoice
